backend/requirements.txt
fastapi
uvicorn
openai-whisper
qdrant-client
sentence-transformers
httpx
